Moved default wrappers into own file
