+++ test/ciphers/feal/integration/test_feal_commands.py
@@ -3,12 +3,7 @@ import unittest
 # noinspection PyUnresolvedReferences
 import test.context
 from ciphers.feal import feal
-from test.sysv_patcher import sysv_patcher
-
-default_encrypt_args = sysv_patcher('feal', 'encrypt', key='0x123456789ABCDEF0123456789ABCDEF', text='0x0')
-
-default_decrypt_args = sysv_patcher('feal', 'decrypt', key='0x123456789ABCDEF0123456789ABCDEF',
-                                    text='0x9C9B54973DF685F8')
+from test.ciphers.feal.integration.wrappers import default_decrypt_args, default_encrypt_args
 
 
 class TestFeal(unittest.TestCase):
